units: introduce getty-pre.target (#6667)
This new target is a passive unit, hence it is supposed to be pulled in
to the transaction by the service that wants to block login on the
console (e.g. text version of initial-setup). Now both getty and
serial-getty are ordered after this target.

https://lists.freedesktop.org/archives/systemd-devel/2015-July/033754.html

(cherry picked from commit 175902541852fb9207f6e532d8da48c9a102340c)

Conflicts:
	units/meson.build

Resolves: #1173080